+++ CHANGELOG.md
@@ -2,7 +2,7 @@ CHANGELOG for Sulu
 ==================
 
 * dev-develop
-    * FEATURE     #1424 [RestComponent]   Implemented and integrated expressions for the listbuilder
+    * FEATURE     #1424 [All]             Implemented and integrated expressions for the listbuilder
     * FEATURE     #1406 [ResourceBundle]  Integrated filters into user settings
     * FEATURE     #1406 [SecurityBundle]  Added api method to delete user settings
     * FEATURE     #1404 [AdminBundle]     Implemented new login design